Move parse_bool_env to picard.env and use it for build variables
Move parse_bool_env() out of picard.util into a new dependency-free
top-level module picard/env.py. Anything under picard.util pulls in
picard/util/__init__.py, which imports PyQt6; picard/__init__.py is
Qt-free, so picard.env can be imported from build tooling (setup.py,
picard.spec) without dragging in Qt.

With that in place, migrate the build-tooling boolean variables to the
shared helper:

- setup.py: PICARD_DISABLE_AUTOUPDATE
- picard.spec: PICARD_BUILD_PORTABLE

These now follow the unified truthiness convention. The values used by CI
(1 or unset) are unaffected; only previously-loose inputs such as '0' or
'false' change, and now evaluate as expected (false) instead of true.

Update the importers (picard.ui, picard.ui.theme), move the tests to
test/test_env.py, and update docs/EnvironmentVariables.md accordingly.
