+++ test/util/test_init.py
@@ -82,7 +82,6 @@ from picard.util import (
     limited_join,
     make_filename_from_title,
     normpath,
-    parse_bool_env,
     parse_date,
     pattern_as_regex,
     resolve_fs_path,
@@ -1212,65 +1211,3 @@ class URLsTest(PicardTestCase):
             key,
             test_text,
         )
-
-
-class ParseBoolEnvTest(PicardTestCase):
-    ENV_NAME = 'PICARD_TEST_BOOL_ENV'
-
-    def _set_env(self, value):
-        if value is None:
-            os.environ.pop(self.ENV_NAME, None)
-        else:
-            os.environ[self.ENV_NAME] = value
-        self.addCleanup(os.environ.pop, self.ENV_NAME, None)
-
-    @subtest_cases(
-        "value",
-        {
-            'one': ('1',),
-            'true': ('true',),
-            'true uppercase': ('TRUE',),
-            'true mixed case': ('True',),
-            'true with whitespace': ('  true  ',),
-            'yes': ('yes',),
-            'on': ('on',),
-        },
-    )
-    def test_truthy_values(self, value):
-        self._set_env(value)
-        self.assertTrue(parse_bool_env(self.ENV_NAME))
-
-    @subtest_cases(
-        "value",
-        {
-            'zero': ('0',),
-            'false': ('false',),
-            'false uppercase': ('FALSE',),
-            'no': ('no',),
-            'off': ('off',),
-            'empty': ('',),
-            'whitespace only': ('   ',),
-        },
-    )
-    def test_falsy_values(self, value):
-        self._set_env(value)
-        self.assertFalse(parse_bool_env(self.ENV_NAME, default=True))
-
-    def test_unset_returns_default(self):
-        self._set_env(None)
-        self.assertFalse(parse_bool_env(self.ENV_NAME))
-        self.assertTrue(parse_bool_env(self.ENV_NAME, default=True))
-
-    @subtest_cases(
-        "value",
-        {
-            'arbitrary word': ('maybe',),
-            'number other than 0/1': ('2',),
-            'partial': ('tru',),
-        },
-    )
-    def test_unrecognized_returns_default(self, value):
-        self._set_env(value)
-        # Unrecognized values fall back to the provided default.
-        self.assertFalse(parse_bool_env(self.ENV_NAME, default=False))
-        self.assertTrue(parse_bool_env(self.ENV_NAME, default=True))
